Support auto param generation logic, validation
Cumulus supports parameter auto-generation.  When you add a resource
that expects a parameter, cumulus will by default add these in for you.

In the case of Stacker, or other apps that might wrap cumulus, there is
a use case of parameters being created on the template for you. To
support this you can turn off "auto parameter generation" in the chain
context.

The behaviour here will not add parameters, but will validate those to
give the consumer a more intelligent message than "resource 'ParamOne'
not found".

Example code to be used in your chain Step can be found in the test.